+++ burnhub-client/src/components/item/Items.js
@@ -1,15 +1,25 @@
-import { useState, useEffect } from "react"
-import { FetchItems } from "../APIManager"
+import { Fragment, useState, useEffect } from "react"
+import { Dialog, Disclosure, Menu, Transition } from '@headlessui/react'
+import { XMarkIcon } from '@heroicons/react/24/outline'
+import { ChevronDownIcon, FunnelIcon, MinusIcon, PlusIcon, Squares2X2Icon } from '@heroicons/react/20/solid'
+import { FetchItems, GetCategories } from "../APIManager"
 import { ItemCard } from "./ItemCard"
 
-export const Items = ({ itemCategory }) => {
+export const Items = () => {
 
     const [items, setItems] = useState([])
     const [filteredItems, setFilteredItems] = useState([])
+    const [categories, setCategories] = useState([])
+    const [category, setCategory] = useState(0)
 
     const fetchItems = async () => {
         const itemsArray = await FetchItems()
         setItems(itemsArray)
+    }
+
+    const fetchCategories = async () => {
+        const categories = await GetCategories()
+        setCategories(categories)
     }
 
     useEffect(() => {
@@ -17,40 +27,277 @@ export const Items = ({ itemCategory }) => {
     }, [])
 
     useEffect(() => {
-        if (itemCategory) {
-            const categoryTickets = items.filter(
-                (item) => item.categoryId === 1
-            );
-            setFilteredItems(categoryTickets);
+        fetchCategories()
+    }, [])
+
+    useEffect(() => {
+        if (category > 0) {
+            const itemsInCategory = items.filter(
+                (item) => item.category.id === category
+            )
+            setFilteredItems(itemsInCategory)
         } else {
-            setFilteredItems(items);
+            setFilteredItems(items)
         }
-    }, [itemCategory])
+    }, [category, items])
+
+    const sortOptions = [
+        { name: 'Most Popular', href: '#', current: true },
+        { name: 'Best Rating', href: '#', current: false },
+        { name: 'Newest', href: '#', current: false },
+        { name: 'Price: Low to High', href: '#', current: false },
+        { name: 'Price: High to Low', href: '#', current: false },
+    ]
+
+    function classNames(...classes) {
+        return classes.filter(Boolean).join(' ')
+    }
+
+    const [mobileFiltersOpen, setMobileFiltersOpen] = useState(false)
 
     return (
+        <div className="bg-white">
+            <div>
+                {/* Mobile filter dialog */}
+                <Transition.Root show={mobileFiltersOpen} as={Fragment}>
+                    <Dialog as="div" className="relative z-40 lg:hidden" onClose={setMobileFiltersOpen}>
+                        <Transition.Child
+                            as={Fragment}
+                            enter="transition-opacity ease-linear duration-300"
+                            enterFrom="opacity-0"
+                            enterTo="opacity-100"
+                            leave="transition-opacity ease-linear duration-300"
+                            leaveFrom="opacity-100"
+                            leaveTo="opacity-0"
+                        >
+                            <div className="fixed inset-0 bg-black bg-opacity-25" />
+                        </Transition.Child>
 
-        <main>
+                        <div className="fixed inset-0 z-40 flex">
+                            <Transition.Child
+                                as={Fragment}
+                                enter="transition ease-in-out duration-300 transform"
+                                enterFrom="translate-x-full"
+                                enterTo="translate-x-0"
+                                leave="transition ease-in-out duration-300 transform"
+                                leaveFrom="translate-x-0"
+                                leaveTo="translate-x-full"
+                            >
+                                <Dialog.Panel className="relative ml-auto flex h-full w-full max-w-xs flex-col overflow-y-auto bg-white py-4 pb-12 shadow-xl">
+                                    <div className="flex items-center justify-between px-4">
+                                        <h2 className="text-lg font-medium text-gray-900">Filters</h2>
+                                        <button
+                                            type="button"
+                                            className="-mr-2 flex h-10 w-10 items-center justify-center rounded-md bg-white p-2 text-gray-400"
+                                            onClick={() => setMobileFiltersOpen(false)}
+                                        >
+                                            <span className="sr-only">Close menu</span>
+                                            <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+                                        </button>
+                                    </div>
 
-            <div className="bg-white">
-                <div className="mx-auto max-w-2xl px-4 py-16 sm:px-6 sm:py-24 lg:max-w-7xl lg:px-8">
-                    <h2 className="sr-only">Products</h2>
+                                    {/* Filters */}
+                                    <form className="mt-4 border-t border-gray-200">
+                                        <h3 className="sr-only">Categories</h3>
+                                        <Disclosure as="div" className="border-t border-gray-200 px-4 py-6">
+                                            {({ open }) => (
+                                                <>
+                                                    <h3 className="-mx-2 -my-3 flow-root">
+                                                        <Disclosure.Button className="flex w-full items-center justify-between bg-white px-2 py-3 text-gray-400 hover:text-gray-500">
+                                                            <span className="font-medium text-gray-900">Category</span>
+                                                            <span className="ml-6 flex items-center">
+                                                                {open ? (
+                                                                    <MinusIcon className="h-5 w-5" aria-hidden="true" />
+                                                                ) : (
+                                                                    <PlusIcon className="h-5 w-5" aria-hidden="true" />
+                                                                )}
+                                                            </span>
+                                                        </Disclosure.Button>
+                                                    </h3>
+                                                    <Disclosure.Panel className="pt-6">
+                                                        <div className="space-y-6">
+                                                            {categories.map((category, categoryIdx) => (
+                                                                <div key={category.id} className="flex items-center">
+                                                                    <input
+                                                                        id={`filter-mobile-${category.id}-${categoryIdx}`}
+                                                                        name={`${category.id}[]`}
+                                                                        defaultValue={category.id}
+                                                                        type="checkbox"
+                                                                        className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                                                                    />
+                                                                    <label
+                                                                        htmlFor={`filter-mobile-${category.id}-${categoryIdx}`}
+                                                                        className="ml-3 min-w-0 flex-1 text-gray-500"
+                                                                    >
+                                                                        {category.name}
+                                                                    </label>
+                                                                </div>
+                                                            ))}
+                                                        </div>
+                                                    </Disclosure.Panel>
+                                                </>
+                                            )}
+                                        </Disclosure>
+                                    </form>
+                                </Dialog.Panel>
+                            </Transition.Child>
+                        </div>
+                    </Dialog>
+                </Transition.Root>
 
-                    <div className="grid grid-cols-1 gap-x-6 gap-y-10 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 xl:gap-x-8">
-                        {filteredItems.map((item) => {
-                            return (
-                                <ItemCard
-                                    itemId={item.id}
-                                    image={item.image}
-                                    name={item.name}
-                                    price={item.price}
-                                />
-                            )
-                        })}
+                <main className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
+                    <div className="flex items-baseline justify-between border-b border-gray-200 pb-6 pt-24">
+                        <h1 className="text-4xl font-bold tracking-tight text-gray-900">Products</h1>
+
+                        <div className="flex items-center">
+                            <Menu as="div" className="relative inline-block text-left">
+                                <div>
+                                    <Menu.Button className="group inline-flex justify-center text-sm font-medium text-gray-700 hover:text-gray-900">
+                                        Sort
+                                        <ChevronDownIcon
+                                            className="-mr-1 ml-1 h-5 w-5 flex-shrink-0 text-gray-400 group-hover:text-gray-500"
+                                            aria-hidden="true"
+                                        />
+                                    </Menu.Button>
+                                </div>
+
+                                <Transition
+                                    as={Fragment}
+                                    enter="transition ease-out duration-100"
+                                    enterFrom="transform opacity-0 scale-95"
+                                    enterTo="transform opacity-100 scale-100"
+                                    leave="transition ease-in duration-75"
+                                    leaveFrom="transform opacity-100 scale-100"
+                                    leaveTo="transform opacity-0 scale-95"
+                                >
+                                    <Menu.Items className="absolute right-0 z-10 mt-2 w-40 origin-top-right rounded-md bg-white shadow-2xl ring-1 ring-black ring-opacity-5 focus:outline-none">
+                                        <div className="py-1">
+                                            {sortOptions.map((option) => (
+                                                <Menu.Item key={option.name}>
+                                                    {({ active }) => (
+                                                        <a
+                                                            href={option.href}
+                                                            className={classNames(
+                                                                option.current ? 'font-medium text-gray-900' : 'text-gray-500',
+                                                                active ? 'bg-gray-100' : '',
+                                                                'block px-4 py-2 text-sm'
+                                                            )}
+                                                        >
+                                                            {option.name}
+                                                        </a>
+                                                    )}
+                                                </Menu.Item>
+                                            ))}
+                                        </div>
+                                    </Menu.Items>
+                                </Transition>
+                            </Menu>
+
+                            <button type="button" className="-m-2 ml-5 p-2 text-gray-400 hover:text-gray-500 sm:ml-7">
+                                <span className="sr-only">View grid</span>
+                                <Squares2X2Icon className="h-5 w-5" aria-hidden="true" />
+                            </button>
+                            <button
+                                type="button"
+                                className="-m-2 ml-4 p-2 text-gray-400 hover:text-gray-500 sm:ml-6 lg:hidden"
+                                onClick={() => setMobileFiltersOpen(true)}
+                            >
+                                <span className="sr-only">Filters</span>
+                                <FunnelIcon className="h-5 w-5" aria-hidden="true" />
+                            </button>
+                        </div>
                     </div>
-                </div>
-            </div>
 
-        </main>
+                    <section aria-labelledby="products-heading" className="pb-24 pt-6">
+                        <h2 id="products-heading" className="sr-only">
+                            Products
+                        </h2>
+
+                        <div className="grid grid-cols-1 gap-x-8 gap-y-10 lg:grid-cols-4">
+                            {/* Filters */}
+                            <form className="hidden lg:block">
+                                <h3 className="sr-only">Categories</h3>
+                                <ul role="list" className="space-y-4 border-b border-gray-200 pb-6 text-sm font-medium text-gray-900">
+                                    {/* {categories.map((category) => (
+                    <>
+                    <span type="button" key={category.id} id={category.id} onClick={(event) => {setCategory(parseInt(event.currentTarget.id))}}>{category.name}</span><br></br>
+                    </>
+                  ))} */}
+                                    <button type="button" onClick={() => { setCategory(0) }}>Reset Filters</button>
+                                </ul>
+
+                                {/* ITEM CATEGORY API */}
+                                <Disclosure as="div" className="border-b border-gray-200 py-6">
+                                    {({ open }) => (
+                                        <>
+                                            <h3 className="-my-3 flow-root">
+                                                <Disclosure.Button className="flex w-full items-center justify-between bg-white py-3 text-sm text-gray-400 hover:text-gray-500">
+                                                    <span className="font-medium text-gray-900">Category</span>
+                                                    <span className="ml-6 flex items-center">
+                                                        {open ? (
+                                                            <MinusIcon className="h-5 w-5" aria-hidden="true" />
+                                                        ) : (
+                                                            <PlusIcon className="h-5 w-5" aria-hidden="true" />
+                                                        )}
+                                                    </span>
+                                                </Disclosure.Button>
+                                            </h3>
+                                            <Disclosure.Panel className="pt-6">
+                                                <div className="space-y-4">
+                                                    {categories.map((category, categoryIdx) => (
+                                                        <div key={category.id} className="flex items-center">
+                                                            <input
+                                                                id={`filter-mobile-${category.id}-${categoryIdx}`}
+                                                                name={`${category.id}[]`}
+                                                                defaultValue={category.id}
+                                                                type="checkbox"
+                                                                className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                                                                onClick={() => { setCategory(category.id) }}
+                                                            />
+                                                            <label
+                                                                htmlFor={`filter-mobile-${category.id}-${categoryIdx}`}
+                                                                className="ml-3 text-sm text-gray-600"
+                                                            >
+                                                                {category.name}
+                                                            </label>
+                                                        </div>
+                                                    ))}
+                                                </div>
+                                            </Disclosure.Panel>
+                                        </>
+                                    )}
+                                </Disclosure>
+                            </form>
+
+                            <div className="lg:col-span-3">
+                                <main>
+
+                                    <div className="bg-white">
+                                        <div className="mx-auto max-w-2xl px-4 py-16 sm:px-6 sm:py-24 lg:max-w-7xl lg:px-8">
+                                            <h2 className="sr-only">Products</h2>
+
+                                            <div className="grid grid-cols-1 gap-x-6 gap-y-10 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 xl:gap-x-8">
+                                                {filteredItems.map((item) => {
+                                                    return (
+                                                        <ItemCard
+                                                            itemId={item.id}
+                                                            image={item.image}
+                                                            name={item.name}
+                                                            price={item.price}
+                                                        />
+                                                    )
+                                                })}
+                                            </div>
+                                        </div>
+                                    </div>
+
+                                </main>
+                            </div>
+                        </div>
+                    </section>
+                </main>
+            </div>
+        </div>
 
     )
 }
