+++ burnhub-client/src/components/views/ApplicationViews.js
@@ -5,7 +5,6 @@ import { ProductForm } from "../productForm/ProductForm";
 import { Items } from "../item/Items";
 import { ItemDetail } from "../item/ItemDetail";
 import { Profile } from "../profile/Profile";
-import { Filter } from "../filters/Filter";
 
 export const ApplicationViews = () => {
 
@@ -19,10 +18,9 @@ export const ApplicationViews = () => {
             }>
             <Route path="/" element={ <Home /> } />
             <Route path="ProductForm" element={<ProductForm />} />
-            <Route path="Items" element={<Filter />} />
+            <Route path="Items" element={<Items />} />
             <Route path="items/:itemId" element={<ItemDetail />} />
             <Route path="/:userId" element={<Profile />} />
-            {/* <Route path="filters" element={<Filter />} /> */}
             {/* <Route path="/store" element={ <Store /> } />
             <Route path="/" element={ <ItemPage /> } />
             <Route path="/" element={ <ViewStores /> } /> */}
